+++ src/clanker_bot/command_handlers/__init__.py
@@ -5,7 +5,7 @@ from .admin import (
     handle_admin_allow_new_meetings,
     handle_admin_stop_new_meetings,
 )
-from .chat import handle_chat, handle_shitpost, handle_speak
+from .chat import handle_chat, handle_shitpost, handle_shitpost_preview, handle_speak
 from .messages import ResponseMessage
 from .types import BotDependencies
 from .voice import handle_join, handle_leave
@@ -20,5 +20,6 @@ __all__ = [
     "handle_join",
     "handle_leave",
     "handle_shitpost",
+    "handle_shitpost_preview",
     "handle_speak",
 ]